Renamed file cxx_polyfill.h to cxx_type_traits_polyfill.h
