+++ dev/ast/where.h
@@ -4,7 +4,7 @@
 #include <utility>  //  std::move
 
 #include "../functional/cxx_universal.h"
-#include "../functional/cxx_polyfill.h"
+#include "../functional/cxx_type_traits_polyfill.h"
 #include "../serialize_result_type.h"
 
 namespace sqlite_orm {
