+++ dev/conditions.h
@@ -7,7 +7,7 @@
 #include <sstream>  //  std::stringstream
 
 #include "functional/cxx_universal.h"
-#include "functional/cxx_polyfill.h"
+#include "functional/cxx_type_traits_polyfill.h"
 #include "type_traits.h"
 #include "collate_argument.h"
 #include "constraints.h"
