+++ dev/functional/cxx_functional_polyfill.h
@@ -6,7 +6,7 @@
 #include <utility>  //  std::forward
 
 #if __cpp_lib_invoke < 201411L
-#include "cxx_polyfill.h"
+#include "cxx_type_traits_polyfill.h"
 #endif
 #include "../member_traits/member_traits.h"
