+++ dev/prepared_statement.h
@@ -8,7 +8,7 @@
 #include <utility>  //  std::pair
 
 #include "functional/cxx_universal.h"
-#include "functional/cxx_polyfill.h"
+#include "functional/cxx_type_traits_polyfill.h"
 #include "functional/cxx_functional_polyfill.h"
 #include "tuple_helper/tuple_filter.h"
 #include "connection_holder.h"
