+++ dev/table_name_collector.h
@@ -5,7 +5,7 @@
 #include <functional>  //  std::function
 #include <typeindex>  //  std::type_index
 
-#include "functional/cxx_polyfill.h"
+#include "functional/cxx_type_traits_polyfill.h"
 #include "type_traits.h"
 #include "select_constraints.h"
 #include "alias.h"
