+++ dev/tuple_helper/tuple_iteration.h
@@ -5,7 +5,7 @@
 #include <utility>  //  std::forward, std::move
 
 #include "../functional/cxx_universal.h"
-#include "../functional/cxx_polyfill.h"
+#include "../functional/cxx_type_traits_polyfill.h"
 #include "../functional/cxx_functional_polyfill.h"
 
 namespace sqlite_orm {
