+++ dev/tuple_helper/tuple_traits.h
@@ -3,7 +3,7 @@
 #include <type_traits>  //  std::is_same
 #include <tuple>
 
-#include "../functional/cxx_polyfill.h"
+#include "../functional/cxx_type_traits_polyfill.h"
 #include "../functional/mpl.h"
 
 namespace sqlite_orm {
