+++ dev/type_printer.h
@@ -5,7 +5,7 @@
 #include <vector>  //  std::vector
 #include "functional/cxx_optional.h"
 
-#include "functional/cxx_polyfill.h"
+#include "functional/cxx_type_traits_polyfill.h"
 #include "type_traits.h"
 #include "is_std_ptr.h"
