+++ dev/xdestroy_handling.h
@@ -6,7 +6,7 @@
 #endif
 
 #include "functional/cxx_universal.h"
-#include "functional/cxx_polyfill.h"
+#include "functional/cxx_type_traits_polyfill.h"
 
 namespace sqlite_orm {
